+++ examples/cuda-flake-parts/flake.nix
@@ -59,6 +59,13 @@
               GALLIUM_DRIVER = "d3d12";
             };
           };
+          buildEnv = [
+            {
+              name = "LD_LIBRARY_PATH";
+              op = "+=(path)";
+              value = "/usr/lib/wsl/lib";
+            }
+          ];
           remotes.local = {
             url = "./repo";
             local = true;